[Major Update] Replaced numPad with the new sprite-based one
Big change made with updating a bunch of things to be class instances rather than variables in create( ) so that they can be accessed by helper methods. More scope errors may be encountered in the future, be sure to class scope them and include .this prefix so that they're accessible throughout the class GameScene.